WIP: destroy mutator + test (pipelines only); Graph is now generic

+++ bundle/phases/destroy.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"net/http"
+	"strings"
 
 	"github.com/databricks/cli/bundle"
 	"github.com/databricks/cli/bundle/deploy/files"
@@ -11,6 +12,7 @@ import (
 	"github.com/databricks/cli/bundle/deploy/terraform"
 	"github.com/databricks/cli/bundle/deployplan"
 	"github.com/databricks/cli/bundle/statemgmt"
+	"github.com/databricks/cli/bundle/terranova"
 	"github.com/databricks/cli/libs/cmdio"
 	"github.com/databricks/cli/libs/diag"
 	"github.com/databricks/cli/libs/log"
@@ -30,18 +32,39 @@ func assertRootPathExists(ctx context.Context, b *bundle.Bundle) (bool, error) {
 	return true, err
 }
 
-func approvalForDestroy(ctx context.Context, b *bundle.Bundle) (bool, error) {
+func getDeleteActions(ctx context.Context, b *bundle.Bundle) ([]terraformlib.Action, error) {
+	if b.DirectDeployment {
+		allResources := b.ResourceDatabase.GetAllResources()
+		var deleteActions []terraformlib.Action
+		for _, node := range allResources {
+			rType, _ := strings.CutSuffix(node.Section, "s")
+			deleteActions = append(deleteActions, terraformlib.Action{
+				Action:       terraformlib.ActionTypeDelete,
+				ResourceType: rType,
+				ResourceName: node.Name,
+			})
+		}
+		return deleteActions, nil
+	}
+
 	tf := b.Terraform
+
 	if tf == nil {
-		return false, errors.New("terraform not initialized")
+		return nil, errors.New("terraform not initialized")
 	}
 
 	actions, err := terraform.ShowPlanFile(ctx, tf, b.Plan.Path)
 	if err != nil {
-		return false, err
+		return nil, err
 	}
 
 	deleteActions := deployplan.Filter(actions, deployplan.ActionTypeDelete)
+
+	return deleteActions, nil
+}
+
+func approvalForDestroy(ctx context.Context, b *bundle.Bundle) (bool, error) {
+	deleteActions, err := getDeleteActions(ctx, b)
 
 	if len(deleteActions) > 0 {
 		cmdio.LogString(ctx, "The following resources will be deleted:")
@@ -68,11 +91,20 @@ func approvalForDestroy(ctx context.Context, b *bundle.Bundle) (bool, error) {
 }
 
 func destroyCore(ctx context.Context, b *bundle.Bundle) diag.Diagnostics {
-	// Core destructive mutators for destroy. These require informed user consent.
-	diags := bundle.ApplySeq(ctx, b,
-		terraform.Apply(),
-		files.Delete(),
-	)
+	var diags diag.Diagnostics
+
+	if b.DirectDeployment {
+		diags = bundle.Apply(ctx, b, terranova.TerranovaDestroy())
+	} else {
+		// Core destructive mutators for destroy. These require informed user consent.
+		diags = bundle.Apply(ctx, b, terraform.Apply())
+	}
+
+	if diags.HasError() {
+		return diags
+	}
+
+	diags = diags.Extend(bundle.Apply(ctx, b, files.Delete()))
 
 	if !diags.HasError() {
 		cmdio.LogString(ctx, "Destroy complete!")
@@ -104,12 +136,14 @@ func Destroy(ctx context.Context, b *bundle.Bundle) (diags diag.Diagnostics) {
 		diags = diags.Extend(bundle.Apply(ctx, b, lock.Release(lock.GoalDestroy)))
 	}()
 
-	diags = diags.Extend(bundle.ApplySeq(ctx, b,
-		statemgmt.StatePull(),
-		terraform.Interpolate(),
-		terraform.Write(),
-		terraform.Plan(terraform.PlanGoal("destroy")),
-	))
+	if !b.DirectDeployment {
+		diags = diags.Extend(bundle.ApplySeq(ctx, b,
+			statemgmt.StatePull(),
+			terraform.Interpolate(),
+			terraform.Write(),
+			terraform.Plan(terraform.PlanGoal("destroy")),
+		))
+	}
 
 	if diags.HasError() {
 		return diags
